flower_frontend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.3.0
numpy==1.26.0
opencv-python==4.8.0.74
torch>=1.8.0
torchvision>=0.9.0
gitpython>=3.1.30
PyYAML>=5.3.1
requests>=2.32.2
scipy>=1.4.1
tqdm>=4.66.3
ultralytics>=8.2.64